requirements.txt
streamlit
torch
torchvision
numpy
pandas
matplotlib
Pillow
scikit-learn